Eidos: fp16 stability fixes — encoder runs fp32, clamped z [-50,50], fp32-safe normalize/norm/cosine/softplus
Root cause: AMP fp16 softmax overflow in DiffAttention (QK^T > 11 -> exp overflow -> NaN cascade).
Encoder now runs fp32 via autocast(enabled=False). Downstream fp16 ops (F.normalize, torch.norm,
F.cosine_similarity, F.softplus) explicitly cast to fp32 before computation.

Changes:
- model.py: encoder wrapped in autocast(enabled=False), z clamp [-50,50], cosine_similarity fp32
- predictive_coder.py: normalize + norm cast to fp32
- neural_memory.py: normalize + norm cast to fp32, NaN guard before read (not after)
- diff_attention.py: removed sdpa_kernel wrapper (encoder fp32 makes it unnecessary)
- ltc_cortex.py: softplus cast to fp32
- mtp_head.py: all-100 labels NaN guard
- synthetic_data.py: chain arithmetic capped, pure multiplication capped
- train.py: accuracy check with v2 regex, every-4-epoch testing, vocab-mismatch pretrain loading
- integration_test.py: 92 tests for all modules

+++ cortex/modules/ltc_cortex.py
@@ -47,7 +47,8 @@ class LTCCortex(nn.Module):
 
     def _ltc_step(self, x: torch.Tensor, h: torch.Tensor) -> torch.Tensor:
         combined = torch.cat([h, x], dim=-1)
-        tau = F.softplus(self.tau_linear(combined)) + 1e-4
+        tau = F.softplus(self.tau_linear(combined).float()) + 1e-4
+        tau = tau.to(h.dtype)
         dh_val = -h / tau + torch.tanh(self.hidden_proj(h) + self.input_proj(x))
         return h + self.dt_per_step * dh_val
